fix(vtd): give VTD's graphd-ensure a 30s budget (was 5s) to kill a hang
VTD adopts-or-spawns vt-graphd as a sibling during its own startup and only
reports healthy after that ensure resolves. The call inherited graphd's 5s
default while electron's `bindVtDaemonForProject` waits 30s for VTD — a budget
inversion: a slow-but-healthy graphd (large parse, slow disk, cold tsx) makes
VTD `die()` at 5s even though electron would have happily waited, leaving the
renderer spinning on "loading workspace". Pass an explicit 30s so VTD only
fails on a genuinely dead graphd. (A second, independent leg of the same
intermittent hang as the graph-db-client request-timeout fix.)

Refactor (behavior-preserving, required to stay under the 500-line file gate
the fix tripped): extract vtd's two location-independent agent-runtime wiring
helpers — `configureAgentRuntimeForVtd` and `buildPublishTerminalRegistryEvent`
— into `src/config/vtdAgentRuntimeWiring.ts`. `resolveVoicetreeCliPackageDir`
stays in the bin (it is `import.meta.url`-relative) and is passed in. vtd.ts
462 lines; tsc clean (the lone `@vt/paths` error is a pre-existing missing
workspace link in this worktree, on an unchanged import).
